+++ lib/src/theme/components/tooltip.dart
@@ -15,6 +15,8 @@ class ShadTooltipTheme {
     this.anchor,
     this.hoverStrategies,
     this.longPressDuration,
+    this.duration,
+    this.reverseDuration,
   });
 
   final bool merge;
@@ -43,6 +45,12 @@ class ShadTooltipTheme {
   /// {@macro ShadTooltip.longPressDuration}
   final Duration? longPressDuration;
 
+  /// {@macro ShadTooltip.duration}
+  final Duration? duration;
+
+  /// {@macro ShadTooltip.reverseDuration}
+  final Duration? reverseDuration;
+
   static ShadTooltipTheme lerp(
     ShadTooltipTheme a,
     ShadTooltipTheme b,
@@ -66,6 +74,12 @@ class ShadTooltipTheme {
           a.longPressDuration != null && b.longPressDuration != null
               ? lerpDuration(a.longPressDuration!, b.longPressDuration!, t)
               : b.longPressDuration,
+      duration: a.duration != null && b.duration != null
+          ? lerpDuration(a.duration!, b.duration!, t)
+          : b.duration,
+      reverseDuration: a.reverseDuration != null && b.reverseDuration != null
+          ? lerpDuration(a.reverseDuration!, b.reverseDuration!, t)
+          : b.reverseDuration,
     );
   }
 
@@ -79,6 +93,8 @@ class ShadTooltipTheme {
     ShadAnchorBase? anchor,
     ShadHoverStrategies? hoverStrategies,
     Duration? longPressDuration,
+    Duration? duration,
+    Duration? reverseDuration,
   }) {
     return ShadTooltipTheme(
       merge: merge ?? this.merge,
@@ -90,6 +106,8 @@ class ShadTooltipTheme {
       anchor: anchor ?? this.anchor,
       hoverStrategies: hoverStrategies ?? this.hoverStrategies,
       longPressDuration: longPressDuration ?? this.longPressDuration,
+      duration: duration ?? this.duration,
+      reverseDuration: reverseDuration ?? this.reverseDuration,
     );
   }
 
@@ -105,6 +123,8 @@ class ShadTooltipTheme {
       anchor: other.anchor,
       hoverStrategies: other.hoverStrategies,
       longPressDuration: other.longPressDuration,
+      duration: other.duration,
+      reverseDuration: other.reverseDuration,
     );
   }
 
@@ -121,7 +141,9 @@ class ShadTooltipTheme {
         other.decoration == decoration &&
         other.anchor == anchor &&
         other.hoverStrategies == hoverStrategies &&
-        other.longPressDuration == longPressDuration;
+        other.longPressDuration == longPressDuration &&
+        other.duration == duration &&
+        other.reverseDuration == reverseDuration;
   }
 
   @override
@@ -134,6 +156,8 @@ class ShadTooltipTheme {
         decoration.hashCode ^
         anchor.hashCode ^
         hoverStrategies.hashCode ^
-        longPressDuration.hashCode;
+        longPressDuration.hashCode ^
+        duration.hashCode ^
+        reverseDuration.hashCode;
   }
 }