+++ lib/src/theme/themes/default_theme_variant.dart
@@ -233,6 +233,8 @@ class ShadDefaultThemeVariant extends ShadThemeVariant {
         preferBelow: false,
         verticalOffset: 24,
       ),
+      duration: Animate.defaultDuration,
+      reverseDuration: Duration.zero,
     );
   }