+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.22.3
 
+- **FIX**: `ShadTooltip` exit animation, add `duration` and `reverseDuration` to it.
 - **FIX**: `ShadMenubar` no longer steals the focus.
 - **FIX**: `ShadButton` stealing focus when pressed.
 - **FEAT**: Add `stylusHandwritingEnabled` to `ShadInput`.